test: TrialCount 테스트 작성
