+++ src/main/java/racingcar/view/InputView/InputView.java
@@ -5,5 +5,5 @@ import java.util.List;
 public interface InputView {
     List<String> inputCarNames();
 
-    int inputTrialCount();
+    String inputTrialCount();
 }
